config: Add github continuous integration testing pipeline
First pipeline that runs just a few tests.
It'll be hard to run all the test cases as there are right now, as the default backend need a special comedi_test driver. However, there are still a few tests that could be added. For now, start small... We'll extend later.