config: update meteor sim
Adjust positions so that the initial position (0,0,0,0,0) happens to be
the SEM position. That's not very realistic, but typically on a real
system, when Odemis start the stage is already in the SEM position.
Also much more handy for testing.

Also update the focus positions to reflect more the actual positions.