gui img: workaround getImage() leaking opened files on wxPython 4.0.7+
A bug was introduced in wxPython 4.0.7 which prevents a stream passed to
wx.Image() from ever being released from memory. The most direct issue
is that the number of files opened would keep increasing, until reaching
the maximum allowed (1024 by default on Ubuntu). The simplest workaround
is to explicitly close the file. The stream is still around, but takes
very little memory.

It's reported here:
https://github.com/wxWidgets/Phoenix/issues/2025