technolution driver: convert physical flyback time to VA
It turned out, that it is quite useful to adjust the flyback time during testing, debugging and also to find causes for the image artifacts. Make it a VA to be easy accessible. In practice later a "good" value will be found and will most likely be the same for all EA systems.