driver andorcam3: handle better when readout rate is not available
The Sona camera has different readout rate for each gain. It looks like
it's possible to select the readout rate, but actually, after setting
the gain, the readout rate is automatically changed.
To properly reflect it in the Odemis component, always first change
the gain before changing the readout rate. This way we detect that the
readout rate has also changed and we display the actual one.