model dataflow/vattribute: don't wait for the other thread in __del__()
In Python 3, __del__() can be called from the garbage collector. In
practice, it's a very common place for it to be called.

While running the garbage collector, other threads cannot run, so it's
not helpful to wait for it to react to the message. It just ends up
making the gc very slow, as it waits for 1s .

Anyway, in general, it's not useful to wait for the other thread to
finish __del__(). So let's just never wait.