xt client: add autoblanking to .blanker VA and add return value to setter of .external VA
The autoblank functionality doesn't work properly with the EnumeratedVA. The start/stop_generate functions check if .blanker.value is None. However, after setting .blanker.value to None, the setter returns True/False, so the automatic blanking/unblanking in the start/stop_generate functions is actually never executed.
There is a way of fixing this by handling the None case in the setter. However, we also update the VA in .updateSettings. Here we would also have to handle the special case and not update the VA.

External VA setter did not return a value.