driver songbird-sim-ccd: reduce exp metadata from 5 to 0.3s
The simcam simulator uses the exposure time as the base for the signal
level. Everything below get's less signal. That meant that to get a good
image we needed a long exposure time. In the test cases, that meant
that either it was fast but the data was almost dark, or it was slow.

Now, even at 0.1 s exposure, the image is good enough for autofocusing.