acq stitching tiledacq: Add focusing_method parameter
Allows to select between autofocus from time to time (as it used to be),
maximum intensity projection (as it used to be if you pass zlevels),
and completely disable it (as it's often useful). It even supports
autofocusing at each tile, but there is probably little use except for
debugging.

In practice, this is useful for the METEOR and ENZEL to always disable
the autofocusing during overview acquistion, even if 1 zlevel is used.