gui cont LocalisationTab: don't save focus ACTIVE_POS as soon as it moves
On the METEOR, it's now allows to access the tab when FM is not active
(just to look at the previous images acquired). However, as the focus
always slightly move, the position is updated, and that cause the
ACTIVE_POS to be updated to the (almost) DEACTIVE_POS.

Anyway, it's not a big deal to go back always to the safe ACTIVE_POS.

Ideally, the chamber tab (acq.move) will be extra clever and
store/restore it, if it's safe.