gui win acquisition OverviewAcquisitionDialog: improve area computation reliability
If the stage was at the wrong place, an error would be reported... and
then the rest would fail with an exception.
=> Now handle properly when area cannot be computed by setting .area to None

Also handle zlevels computation corner cases when it's near the focus
range.

Also rearange a little bit the function organization to make it clearer.

Also update the test cases to not depend on a simulator, and actually
pass. (got broken when the simulator was adjusted)

Also remove some unneeded code.