acq stitching tiledacq: do not increase the horizontalFoV automatically
The idea was to be extra clever and if the user wants a large FoV, we
automatically change the stream horizontalFoV to optimize the
acquisition (ie, less tiles).

However, there might be a reason that the user selected this FoV.
Often because larger doesn't work (eg, image has issue on the sides),
or because the pixel size should be small on purpose (to see details).

No code relied on it. One caller had to do some workaround to avoid this
code. So it just simplifies the code.