Added image translation pre-align and autofocus before ROA acquisition.
Split off calculating the position of the first tile, so it does not get calculated for each field, but only once.

Added test for checking the first position

Ensured that the stage moves to the first field before starting the megafield acquistion, this way the position in the megafield metadata yaml file is correct.