Fastem conf: blank beam before live view and overview image acquistion
In the XTTKDetector we pause the stream in between each acquired image, this already blanks the beam.
We also force blanking and subsequent unblanking of the beam when autoblanking is selected, this caused it to be blanked during acquisition sometimes, mostly notable in the overview image.
In the fastem configuration set the blanking to False for live view and overview imaging, so it's unblanked during acquisition.

Updated test cases for fastem_conf because the value of the blanker changed