Chamber tab: pump/vent button - display state instead of action as label
It is confusing, that when the chamber is pumped, an action "unload" is displayed as button label. It is more intuitive to display the state "loaded" when pumped and "unloaded" when vented. The ebeam button also indicates a state and not an action. The action for both buttons is defined in the corresponding tooltip.