FASTEM: always show ROA size if ROA is selected/active (#2005)
* FASTEM: always show ROA size if ROA is selected/active

Usually the ROA size is only shown when creating an ROA or when resizing it. However, it is convenient to always show the ROA size of the currently active/selected ROA.
The label is now shown in white on grey background, which is slightly transparent to increase readability. White text becomes often unreadable on a grey-scale image.

* The background color in the class Label was hardcoded. Use the value that is actually set on the attribute.
* Adjust the position of the background box so that the text is centered in the box.
* Provide a grey background box for fastem ROAs, so that the ROA size is readable always against gray-scale images. Text color stays white.