FASTEM bug fix overview acquisition in GUI: add callback with correct overview image number
The callback that was added to the future needs to be called with the overview image number. For this purpose a lambda function was used, which did use provide the number as keyword argument. However, when the callback function is executed it will use the value that was last set as "num". Thus, all overview images where stored with the same number label and as a result only one image at a time was displayed in the GUI though all requested overview images were actually acquired.
Use functools.partial as a wrapper function. It allows to pass the callback function and separate keyword arguemnts. With the keyword argument it is made sure that the current overview image number is stored and that the callback is called with the correct number.