gui comp canvas: improve keyboard control
Add support for changing the focus with PageUp/PageDown.

Fix support for zooming in/out with the standard -/= keys next to
"Backspace", when shift is pressed. The numpad +/- keys already worked.

Make pressing shift consistent so that it it's 10x less. (it used to be
sometimes 5x less, 8x less, 10x less).