Fixes for comp canvas tests.
The default merge ratio was changed from 0.3 to 0.5 in commit d7927b253455baab02e4738572473c9e780712b2.
The test cases still assumed the old ratio of 0.3, a fix has been added to explicetely set the merge ratio in the test cases and check for the correct values.

In test_pyramidal_zoom added a precision for checking the mpp and fixed calculating the expected mpp.