model: Added pointSpreadFunctionSize VA to model.DigitalCamera (#1996)
This is equivalent to the standard deviation of the Gaussian approximation of a fluorescence microscope point spread function, measured in no. of pixels of the camera. It's expressed in px, so that it's clear that this represents the size on the image (instead of expressing in meters and having to explain that this is meters taking into account the magnification, or not).

The value can be used as a characteristic size parameter when filtering an image prior to spot detection.

Will be used to clean-up the code in fastem-calibrations:/src/fastem_calibrations/pattern_calibration.py