gui comp DichotomyOverlay: fix error when cursor hovers out of FoV
If the cursor happens to be outside the FoV, quad_hover() returns (None, None).
In Python 2, this was not an issue as it's possible to compare None to
an int. It didn't do the right thing, but it ended up having roughly the
right effect. That's not true for Python 3, which complains loudly about such
comparison.

=> Explicitly check for None.
Drop the comparaison to max_len as anyway it should never happen.