driver andorcam2: don't send old image if quickly stopped/started
During an acquisition, check if it got cancelled at the last moment.
The user could have stopped, changed settings, and started again, while we were
retrieving the image. Let's not send an image with the old
settings in such case.

There was no specific test case, but this used to make
TestCompositedSpectrometer.test_resolution fail.