driver andorcam3: add max_res argument to limit the maximum resolution
For some applications, the chip area is too big. We just don't want the
user to ever have to know the chip is bigger, because outside of the
area, the signal is useless. So allow to clip the maximum resolution, to
behave as if it was the max res of the hardware.