FASTEM overview acq: remove connection of progress of the underlying future to the main future
This ensures a better progress update on the progress bar in the GUI. tiledacq.py does not have a good time estimation yet. It updates the progress based on the original estimation and assumes a fixed duration per tile. After each tile it removes this fixed time per tile from the original estimation. However, if the estimation was wrong, the update is always wrong. Solution is to temporarily remove the connection to the sub-future and just show the progress in the GUI based on the estimated time, which simply counts down for now.