batch future: fix the progress update
When the batch future was used to represent the progress of multiple futures, only for the first future the progress in the overall progress was correctly shown. As soon as the first future was finished, the updated end of the batch future was incorrectly calculated. The same start of the future was used plus adding the remaining time for the not yet finished futures. Thus, the end time of the batch future moved closer to the start in time. As a result, the GUI displayed always "a few seconds left" for the remaining futures that still needed to be executed.