Driver xt_client: the dwell time range is dependent on the horizontalFoV/magnification.
When the horizontalFoV had updated and the dwell time was set to a value lower than possible the following error was raised:
OSError: [Errno -1037828090] INVALID_RANGE_VALUE Set dwell time failed

Therefore update the dwell time range whenever the horizontalFoV is changed. When the dwell time range updates clip the current dwell time value to be within the new range.