simcam: improve blurring of spots
The sigma used for the Gaussian blurring was calculated based on the depth of field which was multiplied by a random factor of 10. This was countertuned by using a very small refractive index below 1, which is not realistic (ri >= 1). Removed the factor 10. Then a timeout error was raised due to too long waiting time for the image. This was caused by the long runtime for the gaussian_filter function. Solution is to reduce the maximum blur allowed to 50 px and additionally use only the square root of this result as sigma.