fastem gui: make list of calibrations an input argument for FASTEMAlignmentController
Thus, we can reuse the alignment controller with a different list of calibrations in other tabs. It will be made use of in the overview tab, where we will add the optical autofocusing.

Additionally, label the panel in the acquisition tab with number 1 as it will be the first one out of 3 steps.