gui win acquisition OverviewAcquisitionDialog: store the full overview file too
We never stored the whole file, so it was loaded in memory only. That
means it was impossible for the user to load it again. Also, for large
overview, that would require keeping everything in memory.

=> Store the overview acquisition in a ome.tiff file, and store the
independent tiles in a sub-folder with the same name + "-tiles" to avoid
too much files around.