gui cont tabs: automatically refit the overview views after adding content
Typically, the overview view is not centered on the new overview, so it
useful to automtically update the overview view to show that new
content.
In the localization tab, we center on the new data (so that it's clear
what is the new data). In the chamber tab, we center on the entire data acquired so far.

Move the fit_to_bbox() method from FastEMAcquisitionCanvas to
DblMicroscopeCanvas, as it can be useful for all of the canvases.