fastem acquisition manager: add inital time estimate to future
There was not inital time estimate on the future before the task was executed. Add it in the same fashion as for the calibration manager. Also move the updating of the time estimate to the to of the run function before the pre-calibrations start. Add also some time estimate for the calibrations.