fastem calibration and acquisition manager: Instance attribute asm_config defined outside __init__
Add argument to init in calibration manager and extend the configure method to handle the case when the attribute is None due to failure as it is still needed in the finally statement to put the system back into a usable state. Similar for acquisition manager when handling the pre calibrations.