gui cont CryoAcquiController: fix zstack using old focus position
The zstack is centered around the current focus position.
However, the focus position was read only at the moment the zstack
settings were changed. So if the user adjusted the focus afterwards,
the zstack was acquired at the wrong levels. This can be especially bad
if the zstack settings are not changed, in which case the focus level
from when the tab was shown are used.

=> Force the zlevels to be recomputed just before starting the
acquisition.