gui fastem cont/comp: move hack to zoom out canvas at init
Now that fit_to_bbox() has been moved to the main class, it's been
"improved" and doesn't fail if the scale would be too small.

However that meant the hack to detect that the resizing failed didn't
detect it anymore. Instead of hacking it even more, move the zoom out
request to the tab controller, which is more in charge of deciding when
the view should be changed any way. It allows to make it a little bit
hacky...