Added MD_FIELD_FREE_SHIFT to configure MS_POS_COR of the scanner based on the status of the immersion mode VA.
MD_POS_COR shifts the single beam image. When immersion is off (field free on) shift the overview image by a calibrated factor with MD_POS_COR, then set back MD_POS_COR to [0, 0] during live stream imaging (immersion on).
This way the overview image is shifted compared to the live stream, since the live stream is somewhat calibrated with the megafield this also calibrates the overview image with the megafield.