plugins: new update_rot plugin to connect ebeam rotation VA to metadata
Currently, when the SEM is scanned via a external card, but the controls
can be read/written via an API, the .rotation VA shows the rotation, but
the data doesn't have this information in the metadata.

That means that the image is always shown as if rotation is 0°. In many
cases that's fine. However, on the SECOM, that's not enough because the
fine alignment overlay looks at .rotation and assumes that MD_ROTATION
will be set. To compensate for it, it sets MD_ROTATION_COR on the
optical images. This causes the optical images to be incorrectly
rotated!

=> copy .rotation as MD_ROTATION & MD_ROTATION_COR. So the SEM image is
not rotated, and the fine alignement is aware of it, so it doesn't
rotate the optical image.