fastem: don't fail acq/align when get asm configuration fails
In theory already the call to retrieve the asm configuration can fail. In that case trying to put the system back into a "good" state will also fail, as the dictionary containing the settings will be None. Make sure that in this case we do not fail but log a warning.
Rename asm_config_orig to asm_config as after calibration it contains original and calibrated values.