Fastem test cases: raise an error when trying to run the test cases with hardware and no backend running.
Previously when trying to run the test cases with TEST_NOHW=False and no backend running it would try to get the different components, and fail with a non descriptive error.
Raise a more informative error when there is no backend running, and thus it is impossible to get the components.