gui cont: move cryo acquired streams control to the streambar controller
Simplifies the tab controller which was doing a lot of different things.

Now there is a dedicated streambar controller for Cryo localization.
In addition to hiding/showing the streams when the view is selected,
it does the same when a feature is selected.

Among other things, it fixes a bug where *every* stream was hidden when changing feature.