bugreporter: better workaround for notification display
Commit 5bd393831f1 ("improve messages after sending the bug"), switched
CallAfter() to CallLater(). However CallLater() needs to be run in the
main GUI thread, otherwise from time to time it causes crashes. In that
case, the notification wouldn't be shown and a popup would indicate
(incorrectly) that the bug report upload failed.

Revert to CallAfter(), but now wait explicitly before ending the app, so
that the notification stays displayed.