init: handle better git error
We use git just to check the version number in case we are running from
the development branch (and so git repo).

If "git describe" returns an error, we used to completely fail.
Now just fallback to the standard way to detect the version number.
It might be not exactly the right version, but it's much better than not
starting at all.

Useful with the changes to git that prevents running it as a different
user. So running git as root (which is the case when staring the
backend) failed if the git repo was owned by the user.