driver technolution: Send custom metadata to the ASM during fastem acquisition.
The ASM accepts a custom data field to add any metadata. It can be useful to know with what exact settings a megafield was acquired, therefore add calibration data as custom metadata.
This will be stored in the mega_field_meta_data.yml file that is stored with each megafield acquisition.
added test cases