gui cont streams: fix display order of HW Detector VAs
HW VAs are mixed for the emitter and detector, but until now we only
used the emitter order. That meant that if the VAs came from the
detector, the order would be roughly random (except for the VAs who
share the same name).
=> use both orders, one after the other. It's not an issue if a VA is
listed several times, the first occurence will be used.