FASTEM: Return resolution and cell size for cell translation calibration
The cell translation calibration is run for a specific set of values for the single field resolution and the cell complete resolution. Thus, for the acquisition it is necessary to also use this settings. As long as adjustable resolution and cell size are not supported in the calibration itself, also do not allow to use different value for the acquisition.