gui conf: fallback to default project dir if impossible to create a new project folder
It can happen that the default project dir cannot be created, for
instance, it got unmounted, or permissions changed. Let's not completely
fail to start the GUI in such case.
=> fall back to the default project directory.

Also, add a "/" after the root project dir because it's actually supposed to
contain the *previous* project directory, and then the parent folder is
used. As it's directly the parent folder, we need to indicate that with
an extra "/".