driver andorcam2: fix minimum resolution now accepted
The cameras may have a minimum of pixels to send (more than 1). For very
tiny resolutions in theory this could be a limit. Until now we
mistakenly applied this minimum both for X and Y. This prevented some
cameras to use full vertical binning, and instead generated very unclear
errors (empty AssertionError).

=> Correctly check the size and, in the very rare case it's actually
limiting, try to increase a bit the resolution.