driver andorcam2: handle hw trigger
Add support for hardware trigger, by providing a special .hardwareTrigger Event.
When the DataFlow is synchronized with this Event, it requests the
camera to use the hardware trigger for frame start.
Provide a hw_trigger_invert argument to control the trigger inversion if
needed.

Also add discardOldFrames and framePeriod for extra control.