driver xt_client: prepareForScan was only added to the FIBScanner and not to the normal Scanner.
This caused the acquisition code in the detector to fail. This commit adds the prepareForScan to the Scanner class, and moves beam blanking to that function, because the FIBScanner does not support beam blanking.

There were also a few test cases which overwrote the blanker VA with a boolean value, which is also fixed.