driver andorcam2: handle change of sync type during acquisition
It's very unlikely, but in theory, it's allowed to change the
synchronization type (eg, None -> SW trigger -> HW trigger) while
the acquisition is running. So far, this was not handled properly.
In particular, if synchronized on HW trigger, and switch to software
trigger, or nothing it'd still wait for the trigger, although it
should start a new acquisition again.