gui model: raise an clear error if ENZEL has no FIB components
With the new alignment tab, the FIB is required.
=> Properly report the components are missing in the YAML file

Also add them to the standard ENZEL simulator. It doesn't yet work
perfectly because the SimSEM has a .shift VA instead of .beamShift VA on
the XT. Probably the former name should be used everywhere.

Take the opportunity to rename the special XT ENZEL sim to a simpler name.