driver scanner: remove private _external and _blanker VAs
These VAs shouldn't be touched. So we shouldn't show them to the user.
All VAs are shown the GUI and CLI, even if they start with a _.
So instead of keeping a link to the original VA, keep a link to the
original component with the VA, to avoid showing the VAs.