acq/gui: Fix position display for ENZEL/METEOR
Do not change the position (mode) while the stage is moving. In
particular:
* The alignment and localization tab don't get temporarily enabled
  half-way through a movement
* The position switch buttons don't get enabled during movement

Also change the METEOR buttons to turn orange during the move, and green
at the end of the move.

Also, on ENZEL, always report THREE_BEAMS instead of IMAGING.

Also, in case of cancellation, don't log an error.

Also clean-ups:
* ENZEL cryo tilt is not needed anymore
* ENZEL IMAGING is not needed anymore
* Don't use internal _state of Future, but just .done() or .running()

In general the change shouldn't have any effect on movement, but only on
the display of the current position.