driver andorcam2: avoid division by zero when switching from large binning to small binning
When switching from large binning and a small cropping area, to a small
binning, the code attempted to keep the area about the same... but
rounded it to 0 px. That caused the resolution to be too small, and the
new code handling it would divide the minimum resolution by the resolution...
which caused a division by zero.

=> clip the resolution to be at least 1 pixel.