gui model meteor: move stage range check
We had requirement that stage.ACTIVE_RANGE is within
stage-bare.IMAGING_RANGE, but that doesn't really make sense as the
stages are in different referentials. In practice, on many systems, that
was fine because the change of referential didn't affect the value too
much, and IMAGING_RANGE can be quite wide.... but on some systems that
just doesn't work.

=> Move the check to when moving the stage to FM. At least, the first
position reached in FM should be a legal position!