odemisd: extend the file mask to include the execute bit
The execute bit on the mask doesn't really mean much for new files.
However, when creating directories, that allows to add files to the
directory... which is a very useful feature.
No driver seemed to depend on this, however on Ubuntu 20.04, some
libraries depend on OpenMPI, which creates some directories in the
background.
This prevented the backend from being started as a normal user on Ubuntu
20.04.