acq test path: fix some test cases
Make assert_pos_as_in_mode() handle the new "syntax" of the modes, which
can include two types of positions (introduced by the EK acquisition).
Also use a single version of this function for all the test cases.

Remove the automatic shutdown of the backend at the end. This is now
handled properly by start_backend().

Drop the check of the grating on spectral mode. It's now left as-is, as
typically any grating is fine (except the mirror, but we allow it anyway).