csv export: fix bug with unknown acquisition type.
In commit 1dc62698 the last line of the export function was adjust and it called the variable acq_types, but this doesn't exist, it should be acq_type. This commit fixes that and adds a test cases to check a value error is raised when an unknown acquisition type is passed.