acq stitching: only ask registrar positions once
Calling .getPositions() actually does the position computation. That's
quite costly, so we should do it only once, not once per tile.

It speeds up the execution, and might also ensure that the dependent
tiles are really at the same place as the lead tile, in case the call
was not always returning the same value.