driver tescan: improve move handling
Safer relative move by computing the target position based on the
position when starting the move, not when requesting the move (which
could be a lot before if another move is running).

After asking to move, first wait a bit, to ensure it doesn't think
the move already finished.

Not sure if any of these will help the fact that sometimes an
(absolute) move request has no effect.