acq path: distinguish "ek" path and "ek-align" paths
For ek-align, we used to just use the "ek" path. However, we want to
force the grating and the filter to be in "good" positions.
That'd be ugly to do in the GUI, so explicitly place it in the OPM.
It will force the axes everytime the stream starts. The GUI still allows
the user to change to different (and odd) positions while playing the
stream.

For the standard EK streams, the GUI takes care of setting the default
position of these streams when it's added by the user (and the user can
later override it).

Also extend the path test cases, to actually test a EK simulator.

Also found a bug in the Sparc EK polarizer simulator.