gui: fix display of colourmap combo-box on Windows
On Windows, Bitmap.CopyFromBuffer() doesn't work. It generates an error
like:
RuntimeError: Failed to gain raw access to bitmap data.

Anyway, there is actually a better function for what we need: FromBuffer().
So let's use it. Actually use the code from gui.util.img, to avoid
duplicating code.

This fixes the display of the colourmaps in Windows.