driver FixedPositionActuator method improvement and addition of testcase
After moving to a position, the actual reported position might not be exactly the position requested. It’s can be because the stage has a quantized position (ie, number of steps) like the TMCM, or the stage has an encoder which keeps updating to positions very near-by.

When asking to move again to the “same” position, the FixedPositionActuator passes on the request. Usually, it has little effect, because it causes a very tiny move, which is either discarded, or very quick. However, in case of anti-backlash under it, there is 50% chance that it it thinks it’s in the “bad” direction, and causes a small step back, followed by a small step forward. That can be quite long.

=> In FixedPositionActuator, after moving to a position, remember where it arrived. Later on, when requested to move to a position, if the dependency is already at either the requested position, or the previous position, don’t do anything.