driver orsay: rename beamShift -> shift
We had already a VA called .shift for the same meaning (eg, in phenom,
and simsem). Later on, came xt_client which introduced beamShift, and we
picked that one up.
However, as it's on the scanner component, typically "shift" is clear
enough. It's also the oldest name. So let's use that one.
That should ensure that all client code can work on any hardware.

(Still need to rename the xt_client VA)