acq test stream: give extra time
On the VM, sometimes the test cases fail. Most probably because we are
not waiting long enough. The time in the Event.wait() can be quite long,
as it's just the maximum time it can wait. As soon as the image is
ready, it'll finish anyway, so it's fine to put a relatively large
number.