acq move: on large Rx rotations, do the Rx rotation far away
Not sure whether this was a typo, but until now, when Rx was large,
we would move Rz... which doesn't help. The idea is that we can move in
X (because it's not really much affected by the Rx, and then move Rx.
Rz has nothing to do there.

In particular, this prevented from going to LOADING position in some
cases, because Rx was so large that moving in Y didn't succeed.