Fix pytest log filter (#2242)
In commit 25fe610fbf532ab5097f91e9d65ec5298a7bbfdf the log filter was updated to include the short test summary. It was only tested on ubuntu 18.04, where the 'short test summary' is at the beginning of the log.
In Ubuntu 20.04 the 'short test summary' is at the end, therefore everything was filtered out except for the 'short test summary'.
Additionally this commit makes sure that when the message is too long it is cut up into 20 complete lines, instead of the first 1000 characters of the log.