driver xt_client: fix xttoolkit image acquisition in the detector
During fastem overview imaging it was noticed that for each tile the image was acquired twice (SDEV-1095) and the stage was still moving during image acquisition (SDEV-873).
The cause of this was that the stop acquiring message was not sent in time. Therefore after acquiring a single image we would start scanning again,
then while it was scanning trigger the stage move. After the stage move there would be a stop and start message queued.
Move the stop message to before canceling scanning to fix this.