runtests update to only show the short test summary info in the test summary file
The test summary used to show anything except the passed test cases. Because there are a lot of warnings and logs the test summary per module was often more than 40 lines long and then the middle part was cut out.
This resulted in some of the failures being cut out of the summary as well. From now on we only show the short test summary and the warnings in the summary file.
This contains one line per failed test case and a list of warnigns at the end of the file. There is no cutting in the summary anymore, so even if there are more than 40 fails,they will still all be shown.
The summary for each test module starts with 'Running ....', with the filepath of the test module on the dots.

The example test output from ubuntu 20.04 started with the wrong line, so that line was removed.