gui util img: make calculate_ticks() symetrical
Whether the range is from 100 -> 1000 or 1000 -> 100, or -1000 -> -100,
the number of ticks and values should be the same (modulo the sign).

Also allow to have ticks near the border, as it's actually quite useful
when reading the graph, and now the code displaying the values do handle
properly the positioning of text next to the border.