fix picoquant_test shutter test cases
TestHH400_Shutters.test_shutters and TestPH300_Shutters.test_shutters were failing when running all the test cases in a class but not when runnning the test standalone.
After unsubscribing in test_acquire_sub it takes a bit of time for the shutters to reset, therefore add a time.sleep at the end of that test case.