driver avantes: fix simulator when start/stopping multiple times
The simulator didn't properly handle stopping. If a measurement was
stopped and just after a new one started, the first one would somehow
"continue"... in a very weird way, as all the rest of the code assumes
that only one measurement is running at a time. It could even happen if
starting a new measurement just after one finished, as happened in the
test case test_acquisition().

=> Stop using a RepeatingTimer, and introduce a dedicated class to
simulate measurement, using a command queue and FSM. It might be
slightly over-kill but works nicely.