DataFlow: in the subscribe method change the debug log to use '%' instead of ','.
By using '%' the listener is evaluated as a string. When running test_weakref in stream_test.py in some cases it happened that the logging call kept a reference to the listener, so the garbage collection did not happen properly.
This fix ensures there is no reference kept by the logging call.