driver xt_client: support updating xtadapter from ubuntu to windows
Overview of the procedure:
xt_client : get the xtadapter version, type, bitness from the running application on Windows PC
xt_client : check if a newer package is available under /usr/share/xtadapter i.e. the directory where the zip files will be made available from the xtadapater debian package
xt_client: if a newer package (zip file) is available transfer it in the form of bytes
xt_client: Done a pop-up starting a newer version in available, please safety restart the xtadapter application on the Windows PC to install it
xt_adapter : from the bytes, create a zipfile in memory and extract all contents of the zipfile in Downloads directory of the Windows PC
xt_adapter : when the user will restart the xtadpater application, a message box pop-us. If the user would like to update the application then the latest package installer will start