Add fastem support for non-rectangular in the acquisition manager. (#2261)
* Add fastem support for non-rectangular in the acquisition manager.

Add the functionality to be able to acquire non-rectangular or rotated rectangular regions of acquisition (ROA's). For now the previous method is still used for acquisition since the GUI still requires compatibility. Two utility functions were added for flood filling a ndarray, which is used for finding the inside of the polygonal ROA's and for finding the bounding box of a set of polygon coordinates.