gui comp buttons: new ProgressRadioButton
Allows to do represent a button which, when pressed down, goes first to
"progress mode", and then indicate the task is complete.
In practice that will be used on the Cryo Chamber tab to switch between
positions. When pressing a button, eg IMAGING, it first turns orange,
and after the move is complete, it becomes green.